Add s3-wagon at slug-compilation-time.
There's an unfortunate chicken-and-egg problem in Leiningen 1.x where
plugins in the dev-deps list cannot get downloaded if the dependencies
need them to be fetched. In the case of private S3 repositories, this
causes issues, so we provide an alternate buildpack setup where the
private S3 wagon is hard-coded into Leiningen's classpath at
slug-compilation time.